fix: 财务精度与数据完整性加固(v4 迁移)+ 分层修复
- 金额列 REAL→INTEGER(分),同步 JSON 保持"元"口径,导入用 Decimal
- v4 迁移:存量金额换算、币种/快照成对归一化、孤儿清理、非空 syncId 去重
- 外键(级联/置空)+ PRAGMA foreign_keys=ON;CHECK 约束;唯一/二级索引
- data 层移除对 cloud 反向依赖,云 DTO 改由 providers barrel 转发
- UI 统一走 data/models 门面;分类树孤儿检测;Android allowBackup=false
- SyncState(device_id, provider_type) 唯一约束;AA JSON 写入守卫
- drift/drift_dev 配套固定 2.34.0,补 drift_schema_v4.json
- 新增 v4 迁移端到端测试;flutter analyze 无告警,全量测试通过

+++ lib/data/models/import_models.dart
@@ -11,6 +11,8 @@
 
 library;
 
+import 'package:decimal/decimal.dart';
+
 /// 导入分类（CSV/云端恢复统一口径）
 class ImportCategory {
   final String name;
@@ -33,7 +35,8 @@ class ImportCategory {
 /// 导入交易数据
 class ImportTransaction {
   final String type; // 全局仅支出模式，固定为 'expense'
-  final double amount;
+  /// 导入金额(元,Decimal 精确解析,杜绝 CSV/JSON 解析阶段的浮点尾差)。
+  final Decimal amount;
   final String? categoryName;
   final String? categoryKind;
   final DateTime happenedAt;
@@ -48,7 +51,7 @@ class ImportTransaction {
   /// 仅云端全量恢复路径携带:/sync/full 的 tx item 输出的 nativeAmount 是
   /// 源设备记账时按当时汇率折算的真实所见金额。null(CSV 导入/无该字段的
   /// 备份) → 落库时按本地有效汇率重算,与常规导入语义一致。
-  final double? nativeAmount;
+  final Decimal? nativeAmount;
 
   /// 不计入支出统计标记。null(JSON 无此键) → 落库默认 false,
   /// 与 server snapshot「缺键 = false」语义对齐。
